+++ core-ng/src/main/java/core/framework/impl/kafka/KafkaHeaders.java
@@ -1,0 +1,11 @@
+package core.framework.impl.kafka;
+
+/**
+ * @author neo
+ */
+class KafkaHeaders {
+    static final String HEADER_REF_ID = "refId";
+    static final String HEADER_CLIENT = "client";
+    static final String HEADER_CLIENT_IP = "clientIP";
+    static final String HEADER_TRACE = "trace";
+}