+++ core-ng/src/main/java/core/framework/impl/kafka/KafkaMessagePublisher.java
@@ -2,20 +2,18 @@ package core.framework.impl.kafka;
 
 import core.framework.api.kafka.MessagePublisher;
 import core.framework.api.log.ActionLogContext;
-import core.framework.api.util.Maps;
 import core.framework.api.util.Network;
 import core.framework.api.util.StopWatch;
-import core.framework.api.util.Types;
+import core.framework.api.util.Strings;
 import core.framework.impl.json.JSONWriter;
 import core.framework.impl.log.ActionLog;
 import core.framework.impl.log.LogManager;
 import core.framework.impl.log.LogParam;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.header.Headers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.Map;
 
 /**
  * @author neo
@@ -27,14 +25,14 @@ public class KafkaMessagePublisher<T> implements MessagePublisher<T> {
     private final MessageValidator validator;
     private final String topic;
     private final LogManager logManager;
-    private final JSONWriter<KafkaMessage<T>> writer;
+    private final JSONWriter<T> writer;
 
     public KafkaMessagePublisher(Producer<String, byte[]> producer, MessageValidator validator, String topic, Class<T> messageClass, LogManager logManager) {
         this.producer = producer;
         this.validator = validator;
         this.topic = topic;
         this.logManager = logManager;
-        writer = JSONWriter.of(Types.generic(KafkaMessage.class, messageClass));
+        writer = JSONWriter.of(messageClass);
     }
 
     @Override
@@ -48,17 +46,17 @@ public class KafkaMessagePublisher<T> implements MessagePublisher<T> {
 
         StopWatch watch = new StopWatch();
         try {
-            KafkaMessage<T> kafkaMessage = new KafkaMessage<>();
-            Map<String, String> headers = Maps.newHashMap();
-            headers.put(KafkaMessage.HEADER_CLIENT_IP, Network.localHostAddress());
-            headers.put(KafkaMessage.HEADER_CLIENT, logManager.appName);
+            byte[] message = writer.toJSON(value);
+
+            ProducerRecord<String, byte[]> record = new ProducerRecord<>(topic, key, message);
+            Headers headers = record.headers();
+            headers.add(KafkaHeaders.HEADER_CLIENT_IP, Strings.bytes(Network.localHostAddress()));
+            if (logManager.appName != null)
+                headers.add(KafkaHeaders.HEADER_CLIENT, Strings.bytes(logManager.appName));
             linkContext(headers);
-            kafkaMessage.headers = headers;
-            kafkaMessage.value = value;
-            byte[] message = writer.toJSON(kafkaMessage);
 
             logger.debug("publish, topic={}, key={}, message={}", topic, key, LogParam.of(message));
-            producer.send(new ProducerRecord<>(topic, key, message));
+            producer.send(record);
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("kafka", elapsedTime);   // kafka producer send message in background, the main purpose of track is to count how many message sent in action
@@ -66,11 +64,11 @@ public class KafkaMessagePublisher<T> implements MessagePublisher<T> {
         }
     }
 
-    private void linkContext(Map<String, String> headers) {
+    private void linkContext(Headers headers) {
         ActionLog actionLog = logManager.currentActionLog();
         if (actionLog == null) return;
 
-        headers.put(KafkaMessage.HEADER_REF_ID, actionLog.refId());
-        if (actionLog.trace) headers.put(KafkaMessage.HEADER_TRACE, "true");
+        headers.add(KafkaHeaders.HEADER_REF_ID, Strings.bytes(actionLog.refId()));
+        if (actionLog.trace) headers.add(KafkaHeaders.HEADER_TRACE, Strings.bytes("true"));
     }
 }